requirements.txt
requests==2.31.0
python-nmap==0.7.1
python-masscan==1.0.0
impacket==0.11.0
argparse==1.4.0
colorama==0.4.6
pymetasploit3==1.0.3
python-libnmap==0.7.3
scapy==2.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
redis==4.6.0
cryptography==42.0.1
paramiko==3.3.1
schedule==1.2.0
PyYAML==6.0.1
beautifulsoup4==4.12.2
numpy==1.24.3
pandas==2.0.3